Reset styles to match @thefoxis' initial sketch
The point of this repo is to take the PNG that @thefoxis posted at:

https://github.com/gittip/building.gittip.com/issues/16#issuecomment-37030463

And to iterate on it in code. This commit makes the page look pretty
nearly like that initial sketch. I left off the "New / Top / Top"
section because I want to move away from that entirely, and I didn't
bother to make it absolutely pixel-perfect because it was a sketch in
the first place.